requirements.txt
Flask==3.0.0
plaid-python==16.0.0
python-dotenv==1.0.0
